Add set_focus method, use it to set search input field on load
the search page loads without focus on the input text field. Added
a method in form called set focus that will add a line of jquery to
call .focus() on the specified element by id. Added a call in main
to set focus on the text input field for the search.