Update search type selector to default to "Threads" no more "Choose"
the search form loaded up and there was an initial value of "Choose"
for the search type. Removed the initial value and updated the select
box generating code to take a parameter for the option you want to
be SELECTED. Updated the search page to use the new param to select
Threads.

Hopefully the new OR condition still preserves the correct behavior
of a form loading from DB values that set the state.