Added in a parse method to create soundcloud embeds from a song url
soundcloud support was missing from the version of pgBoard that
was used for the intial commit to github.  This method and the
associated calls to it adds that support back in.

This uses the old style flash embed and song URL method of embedding
soundcloud's new embed snippets use a newer style flash player or
html 5 player. Both examples don't use the song url. At some point
we should try to update to the html 5 player so iOS users can see
them too