Strip excessive Unicode combining marks in chat
Application of many Unicode combiners render annoyingly
(incorrectly?) in common browsers and OSes, and this
fact is abused.  The following should limit each normal
character to a single combining mark and alleviate the
problem without stopping many legitimate uses of
combining marks.

Doing it on the browser for now so that the server isn't
burdened with it.

It's possible this should be done for all user-entered
text, though as far as I know it has only been an
issue in chat so far.